DriveStraightCmd fixes.  PIDDrive will not reverse motors at setpoint.
For DriveStraight, fixes bugs that may have caused rough state
transitions (speed changes).  Also, the "backward" and "forward" cases
has been folded into one case to make future algorithm changes/fixes
easier.

For PIDDrive, the output range has been changed to 0-1.0 for the forward
case and -1.0-0 for the backward case.  This should let the motors coast
when they hit the target speed instead of "braking" to maintain the
speed.  In the future, might be able to use the 'F' PID parameter to
help with this.